requirements.txt
django-workflows
